requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
google-genai
chromadb
text-chunker
nltk
semchunk 
tiktoken 
transformers
